split stages
