+++ dashboard/pages/data_projection_screening/stages/__init__.py
@@ -1,14 +1,9 @@
 from .s1_projection_input import PROJECTION_INPUT_STAGE
 from .s2_projection_display import PROJECTION_DISPLAY_STAGE
 from .s3_report import REPORT_STAGE
-from .s4_smiles_input import SMILES_INPUT_STAGE
-from .s5_smiles_display import SMILES_PROJECTION_DISPLAY_STAGE
-
 
 STAGES = [
     PROJECTION_INPUT_STAGE,
     PROJECTION_DISPLAY_STAGE,
     REPORT_STAGE,
-    SMILES_INPUT_STAGE,
-    SMILES_PROJECTION_DISPLAY_STAGE,
 ]
