+++ dashboard/pages/data_projection_screening/stages/s2_projection_display.py
