+++ dashboard/pages/data_projection_smiles/stages/s2_smiles_display.py
